AgroVerse/requirements.txt
Flask==3.1.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.68.1
grpcio-status==1.68.1
gTTS==2.5.4
numpy==2.0.2
pydantic==2.10.4
requests==2.32.3
scikit-learn==1.6.0
scipy==1.13.1
SpeechRecognition==3.12.0
Werkzeug==3.1.3
